Move watchers to cacher
