+++ src/main.ts
@@ -415,7 +415,6 @@ export class Extension {
 
     async dispose() {
         await this.cacher.dispose()
-        await this.manager.dispose()
         this.server.dispose()
         await this.pegParser.dispose()
         await this.mathPreview.dispose()
